function_app/restart_vm/tests/requirements-dev.txt
pytest
azure-functions
pytest-mock